refactor: move `ResponseStatus` and `ResponseStatusFilter` to `enums.py`
